cp hoard.proto into hoard-js
