+++ hoard-js/Makefile
@@ -2,3 +2,4 @@
 protobuf:
 	@grpc_tools_node_protoc --js_out=import_style=commonjs,binary:protobuf --grpc_out=protobuf \
 	--plugin=protoc-gen-grpc=`which grpc_tools_node_protoc_plugin` --proto_path=../protobuf/ hoard.proto
+	@cp ../protobuf/hoard.proto ./protobuf/hoard.proto
