+++ hoard-js/protobuf/hoard.proto
@@ -1,0 +1,103 @@
+syntax = "proto3";
+
+import "github.com/gogo/protobuf/gogoproto/gogo.proto";
+
+import "reference.proto";
+import "grant.proto";
+import "storage.proto";
+
+package hoard;
+
+option go_package = "github.com/monax/hoard";
+
+option (gogoproto.protosizer_all) = true;
+
+service Grant {
+    // Seal a Reference to create a Grant
+    rpc Seal (ReferenceAndGrantSpec) returns (grant.Grant);
+
+    // Unseal a Grant to recover the Reference
+    rpc Unseal (grant.Grant) returns (reference.Ref);
+
+    // Convert one grant to another grant to re-share with another party or just
+    // to change grant type
+    rpc Reseal (GrantAndGrantSpec) returns (grant.Grant);
+
+    // Put a Plaintext and returned the sealed Reference as a Grant
+    rpc PutSeal (PlaintextAndGrantSpec) returns (grant.Grant);
+
+    // Unseal a Grant and follow the Reference to return a Plaintext
+    rpc UnsealGet (grant.Grant) returns (Plaintext);
+}
+
+// Provide plaintext and get plaintext back
+service Cleartext {
+    // Push some plaintext data into storage and get its deterministically
+    // generated secret reference.
+    rpc Put (Plaintext) returns (reference.Ref);
+
+    // Provide a secret reference to an encrypted blob and get the plaintext
+    // data back.
+    rpc Get (reference.Ref) returns (Plaintext);
+}
+
+// Deterministic encryption
+service Encryption {
+    // Encrypt some data and get its deterministically generated
+    // secret reference including its address without storing the data.
+    rpc Encrypt (Plaintext) returns (ReferenceAndCiphertext);
+
+    // Decrypt the provided data by supplying it alongside its secret
+    // reference. The address is not used for decryption and may be omitted.
+    rpc Decrypt (ReferenceAndCiphertext) returns (Plaintext);
+}
+
+// Interact directly with storage backend
+service Storage {
+    // Insert the (presumably) encrypted data provided and get the its address.
+    rpc Push (Ciphertext) returns (Address);
+
+    // Retrieve the (presumably) encrypted data stored at address.
+    rpc Pull (Address) returns (Ciphertext);
+
+    // Get some information about the encrypted blob stored at an address,
+    // including whether it exists.
+    rpc Stat (Address) returns (storage.StatInfo);
+}
+
+message GrantAndGrantSpec {
+    grant.Grant Grant = 1;
+    // The type of grant to output
+    grant.Spec GrantSpec = 2;
+}
+
+message PlaintextAndGrantSpec {
+    Plaintext Plaintext = 1;
+    // The type of grant to output
+    grant.Spec GrantSpec = 2;
+}
+
+message ReferenceAndGrantSpec {
+    reference.Ref Reference = 1;
+    // The type of grant to output
+    grant.Spec GrantSpec = 2;
+}
+
+message Plaintext {
+    bytes Data = 1;
+    bytes Salt = 2;
+}
+
+message Ciphertext {
+    bytes EncryptedData = 1;
+}
+
+message ReferenceAndCiphertext {
+    reference.Ref Reference = 1;
+    Ciphertext Ciphertext = 2;
+}
+
+message Address {
+    bytes Address = 1;
+}
+
